requirements.txt
gunicorn
numpy
pillow
opencv-python-headless
flask
flask_cors
torch
torchvision
requests
